pyproject.toml
[tool.poetry]
name = "yandex-test"
version = "0.1.0"
description = "Тестовое задание яндекса"
authors = ["Romanov Aleksey Leonidovich <romanovaleks@ozon.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
# API libraries
requests = "^2.28.1"
# test runner and plugins
pytest = "^7.1.3"
allure-pytest = "^2.11.1"
pytest-xdist = "^2.5.0"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
# git hooks
pre-commit = "^2.14.0"
# formatters
isort = "^5.9.3"
black = "^22.1.0"

[tool.isort]
profile = "black"
line_length = 110
multi_line_output = 3
include_trailing_comma = true
skip = [".git", ".pytest_cache", ".idea", "__pycache__", ".venv", "venv",]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
